Move the delete button to a module and add flash
