+++ src/delete_button.js
@@ -1,0 +1,14 @@
+import useStorage from './use-storage';
+
+const deleteEvent = (todoList) => {
+  const deleteBtn = document.querySelectorAll('.delete-btn');
+  deleteBtn.forEach(btn => btn.addEventListener('click', (e) => {
+    const parent = e.target.parentNode;
+
+    todoList.splice(parent.id, 1);
+    useStorage('todoList', todoList);
+    parent.remove();
+  }));
+};
+
+export { deleteEvent as default };
